# Determine how to run listeners at setup time instead of execution time